build: rename project
