+++ README.md
@@ -1,0 +1,4 @@
+# Word prediction in Sparv
+
+[![CI](https://github.com/spraakbanken/sparv-word-prediction-plugin/actions/workflows/test.yml/badge.svg)](https://github.com/spraakbanken/sparv-word-prediction-plugin/actions/workflows/test.yml)
+[![Codecov](https://codecov.io/gh/spraakbanken/sparv-word-prediction-plugin/coverage.svg)](https://codecov.io/gh/spraakbanken/sparv-word-prediction-plugin)
