+++ word-prediction--kb-bert/pyproject.toml
@@ -34,10 +34,10 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project.entry-points."sparv.plugin"]
-word_prediction = "word_prediction"
+word_prediction_kb_bert = "word_prediction_kb_bert"
 
 [project.urls]
-Homepage = "https://github.com/spraakbanken/sparv-word-prediction-plugin"
+Homepage = "https://spraakbanken.gu.se"
 Repository = "https://github.com/spraakbanken/sparv-word-prediction-plugin"
 "Bug Tracker" = "https://github.com/spraakbanken/sparv-word-prediction-plugin/issues"
 
@@ -45,7 +45,7 @@ Repository = "https://github.com/spraakbanken/sparv-word-prediction-plugin"
 exclude = ["/.github", "/docs"]
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/word_prediction"]
+packages = ["src/word_prediction_kb_bert"]
 
 [tool.hatch.metadata]
 allow-direct-references = true
