+++ word-prediction--kb-bert/tests/test_predictor.py
@@ -1,6 +1,6 @@
 from typing import Tuple
 import pytest
-from word_prediction.kb_bert.predictor import TopKPredictor
+from word_prediction_kb_bert.predictor import TopKPredictor
 
 
 TEXTS: dict[str, str] = {
